restructured risk samples
